split codes to file
