+++ jwt/jwk.go
@@ -1,0 +1,23 @@
+package jwt
+
+import "gopkg.in/square/go-jose.v2"
+
+type JsonWebKey struct {
+	jose.JSONWebKey
+}
+
+func (key *JsonWebKey) Algorithm() jose.SignatureAlgorithm {
+	return jose.SignatureAlgorithm(key.JSONWebKey.Algorithm)
+}
+
+func (key *JsonWebKey) Use() string {
+	return key.JSONWebKey.Use
+}
+
+func (key *JsonWebKey) Key() interface{} {
+	return key.JSONWebKey.Key
+}
+
+func (key *JsonWebKey) ID() string {
+	return key.JSONWebKey.KeyID
+}
